Refactor: Move AWS WebSocket manager to providers directory
