+++ internal/providers/aws/websocket/manager_test.go
